requirements.txt
beautifulsoup4==4.9.1
discord.py==1.4.1
hjson==3.0.1
requests==2.22.0
youtube-search-python==1.2.8
praw==7.1.0
PyNaCl==1.4.0
youtube-dl==2020.7.28
gitdb==4.0.5
python-git==2018.2.1
